requirements.txt
aiogram==3.15.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
structlog==24.4.0
redis==5.2.1
httpx==0.28.1
python-barcode==0.15.1
Pillow==11.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
